separate transformer store from lib/
